requirements.txt
Django~=4.2
python-decouple~=3.8
drf-yasg~=1.21.5
djangorestframework~=3.14.0
opencv-python~=4.7.0.72
django-cors-headers~=3.14.0
numpy~=1.24.2
imutils~=0.5.4
deepmerge>=1.0.1
dotmap>=1.3.30
jsonschema>=4.16.0
matplotlib>=3.5.3
numpy>=1.23.2
pandas>=1.4.3
rich>=12.6.0
screeninfo>=0.8.1